Various changes for handling the lifecycle of TodoLists including their persistence, caching, and notifications.  Performance improvements.
• Made the PersistenceManager code less dependent on the
ModelManagerService code to create a cleaner separation of concerns.
Some more work still needs to be done.
• Added the ACCESS_LOCATION parameter to a few more PersistenceManager
methods to provide better control to the user of those methods.
• More clearly implemented the definition and the use of the local cache
of TodoLists (see cachedTodoLists in ModelManagerService).
• Made the app startup more efficient by avoiding costly database trips
(local or cloud) because of the availability and usage of the locally
cached TodoLists.
• To track the persistence (creation, modification, and deletion) of a
TodoList and take appropriate action in the UI, created the
PeristenceCallback interface.  Implemented the interface in
ListsViewerActivity and used it in ItemsViewerActivity and
ItemsViewerFragment.  This new approach avoids the use of expensive
refreshes of all the TodoLists after a particular TodoList is created,
modified, or deleted. This change seems to have provided huge
performance improvements.
• Provided useful Toast messages after a return from
ItemsViewerActivity.
• Made many changes to avoid potential null pointer exceptions.
• Removed some unused code.